feat: add statistic
