+++ src/pages/SprintGamePage/components/SprintWindowGame/sprintWindowGame.ts
@@ -2,6 +2,7 @@ import getAllWords, {
   getAggregatedWords,
 } from '../../../../api/Words/WordsAPI';
 import handleProgress from '../../../../components/progress/progress';
+import learnedWord, { setUnlearnedStatusWord } from '../../../../components/studied-words/learned';
 import { generalState } from '../../../../states/generalState';
 import IWordCard from '../../../../types/interfaces/words';
 import dicAndBookVars from '../../../DictionaryBookPages';
@@ -248,6 +249,7 @@ function handleTrueAnswer() {
       sprintGameState.currentWord.id as string || sprintGameState.currentWord.id as string,
       true
     );
+    learnedWord(sprintGameState.currentWord.id as string || sprintGameState.currentWord.id as string, 'sprint-game');
   } else {
     sprintGameState.falseData.push(sprintGameState.currentWord);
     isRight = false;
@@ -275,6 +277,7 @@ function handleFalseAnswer() {
       sprintGameState.currentWord.id as string || sprintGameState.currentWord.id as string,
       false
     );
+    setUnlearnedStatusWord(sprintGameState.currentWord.id as string || sprintGameState.currentWord.id as string, 'sprint-game');
   } else {
     isRight = false;
     sprintGameState.falseData.push(sprintGameState.currentWord);
